vr/sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vrdev"
version = "1.0.0"
description = "Verifiable Rewards for Real-World AI Agent Tasks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "vr.dev" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "httpx>=0.27",
    "jsonschema>=4.0",
    "pyyaml>=6.0",
    'tomli>=2.0; python_version < "3.11"',
]

[project.urls]
Homepage = "https://vr.dev"
Repository = "https://github.com/vrDotDev/vr-dev"
Documentation = "https://vr.dev/docs"
Issues = "https://github.com/vrDotDev/vr-dev/issues"

[project.optional-dependencies]
llm = ["openai>=1.0"]
mcp = ["mcp>=1.26"]
browser = ["playwright>=1.40"]
vision = ["Pillow>=10.0"]
langchain = ["langchain-core>=0.2"]
all = ["openai>=1.0", "mcp>=1.26", "playwright>=1.40", "Pillow>=10.0", "langchain-core>=0.2"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.4",
]

[project.scripts]
vr = "vrdev.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/vrdev"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks integration tests (deselect with '-m not integration')",
]

[tool.coverage.report]
exclude_also = [
    "if __name__",
    "pragma: no cover",
    "except ImportError",
]

[tool.ruff]
target-version = "py310"
line-length = 100
